test: clean up tests
